Animation now works.
Need to work on updating animation frames to get a good rate of
animation. (also potentially interpolation between frames).

Created an animated entity class which gives allows positioning of an
animated model in the world. Also updated animated vertex shader to
allow for a model transformation matrix to move the model around.